move location code to the new module
