--- urbanairship-location-react-native/android/src/main/java/com/urbanairship/reactnative/location/AirshipLocationModule.java
@@ -1,22 +1,0 @@
-package com.urbanairship.reactnative.location;
-
-import com.facebook.react.bridge.ReactApplicationContext;
-import com.facebook.react.bridge.ReactContextBaseJavaModule;
-import com.facebook.react.bridge.ReactMethod;
-import com.facebook.react.bridge.Callback;
-
-public class AirshipLocationModule extends ReactContextBaseJavaModule {
-
-    private final ReactApplicationContext reactContext;
-
-    public AirshipLocationModule(ReactApplicationContext reactContext) {
-        super(reactContext);
-        this.reactContext = reactContext;
-    }
-
-    @Override
-    public String getName() {
-        return "AirshipLocation";
-    }
-
-}
