+++ urbanairship-location-react-native/ios/AirshipLocationReactModule.h
@@ -1,0 +1,14 @@
+/* Copyright Airship and Contributors */
+
+#if __has_include("AirshipLib.h")
+#import "AirshipLib.h"
+#else
+@import Airship;
+#endif
+
+/**
+ * React module for Airship Location
+ */
+@interface AirshipLocationReactModule : NSObject <RCTBridgeModule>
+
+@end
