+++ urbanairship-react-native/ios/UARCTModule/UrbanAirshipReactModule.m
@@ -174,26 +174,6 @@ RCT_REMAP_METHOD(associateIdentifier,
     [[UAirship shared].analytics associateDeviceIdentifiers:identifiers];
 }
 
-RCT_EXPORT_METHOD(setLocationEnabled:(BOOL)enabled) {
-    [UAirship shared].locationProvider.locationUpdatesEnabled = enabled;
-}
-
-RCT_REMAP_METHOD(isLocationEnabled,
-                 isLocationEnabled_resolver:(RCTPromiseResolveBlock)resolve
-                 rejecter:(RCTPromiseRejectBlock)reject) {
-    resolve(@([UAirship shared].locationProvider.isLocationUpdatesEnabled));
-}
-
-RCT_REMAP_METHOD(isBackgroundLocationAllowed,
-                 isBackgroundLocationAllowed_resolver:(RCTPromiseResolveBlock)resolve
-                 rejecter:(RCTPromiseRejectBlock)reject) {
-    resolve(@([UAirship shared].locationProvider.isBackgroundLocationUpdatesAllowed));
-}
-
-RCT_EXPORT_METHOD(setBackgroundLocationAllowed:(BOOL)enabled) {
-    [UAirship shared].locationProvider.backgroundLocationUpdatesAllowed = enabled;
-}
-
 RCT_REMAP_METHOD(runAction,
                  name:(NSString *)name
                  value:(id)value
